cleanup code base
