--- contracts/examples/ERC20Verifier.sol
@@ -1,61 +1,0 @@
-// SPDX-License-Identifier: MIT
-pragma solidity 0.8.16;
-
-import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
-import "../lib/GenesisUtils.sol";
-import "../interfaces/IERC20zkp.sol";
-import "../interfaces/ICircuitValidator.sol";
-import "../ZKPVerifier.sol";
-
-contract ERC20Verifier is ERC20, ZKPVerifier {
-    uint64 public constant TRANSFER_REQUEST_ID = 1;
-
-    mapping(uint256 => address) public idToAddress;
-    mapping(address => uint256) public addressToId;
-
-    uint256 public TOKEN_AMOUNT_FOR_AIRDROP_PER_ID = 5 * 10 ** uint256(decimals());
-
-    constructor(string memory name_, string memory symbol_) ERC20(name_, symbol_) {}
-
-    function _beforeProofSubmit(
-        uint64 /* requestId */,
-        uint256[] memory inputs,
-        ICircuitValidator validator
-    ) internal view override {
-        // check that  challenge input is address of sender
-        address addr = GenesisUtils.int256ToAddress(inputs[validator.getChallengeInputIndex()]);
-        // this is linking between msg.sender and
-        require(_msgSender() == addr, "address in proof is not a sender address");
-    }
-
-    function _afterProofSubmit(
-        uint64 requestId,
-        uint256[] memory inputs,
-        ICircuitValidator validator
-    ) internal override {
-        require(
-            requestId == TRANSFER_REQUEST_ID && addressToId[_msgSender()] == 0,
-            "proof can not be submitted more than once"
-        );
-
-        // address didn't get airdrop tokens
-        uint256 id = inputs[validator.getChallengeInputIndex()];
-        // additional check didn't get airdrop tokens before
-        if (idToAddress[id] == address(0)) {
-            super._mint(_msgSender(), TOKEN_AMOUNT_FOR_AIRDROP_PER_ID);
-            addressToId[_msgSender()] = id;
-            idToAddress[id] = _msgSender();
-        }
-    }
-
-    function _beforeTokenTransfer(
-        address /* from */,
-        address to,
-        uint256 /* amount */
-    ) internal view override {
-        require(
-            proofs[to][TRANSFER_REQUEST_ID],
-            "only identities who provided proof are allowed to receive tokens"
-        );
-    }
-}
