--- contracts/interfaces/IZKPAirdrop.sol
@@ -1,12 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0
-
-pragma solidity 0.8.16;
-
-interface IZKPAirdrop {
-    function mintWithProof(
-        uint256[] memory inputs,
-        uint256[2] memory a,
-        uint256[2][2] memory b,
-        uint256[2] memory c
-    ) external returns (bool);
-}
