--- contracts/lib/verifier.sol
@@ -1,146 +1,0 @@
-//
-// Copyright 2017 Christian Reitwiessner
-// Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
-// The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-//
-// 2019 OKIMS
-//      ported to solidity 0.6
-//      fixed linter warnings
-//      added requiere error messages
-//
-//
-// SPDX-License-Identifier: GPL-3.0
-pragma solidity 0.8.16;
-
-import "../lib/Pairing.sol";
-import "../interfaces/IStateTransitionVerifier.sol";
-
-contract Verifier is IStateTransitionVerifier {
-    using Pairing for *;
-    struct VerifyingKey {
-        Pairing.G1Point alfa1;
-        Pairing.G2Point beta2;
-        Pairing.G2Point gamma2;
-        Pairing.G2Point delta2;
-        Pairing.G1Point[] IC;
-    }
-    struct Proof {
-        Pairing.G1Point A;
-        Pairing.G2Point B;
-        Pairing.G1Point C;
-    }
-
-    function verifyingKey() internal pure returns (VerifyingKey memory vk) {
-        vk.alfa1 = Pairing.G1Point(
-            20491192805390485299153009773594534940189261866228447918068658471970481763042,
-            9383485363053290200918347156157836566562967994039712273449902621266178545958
-        );
-
-        vk.beta2 = Pairing.G2Point(
-            [
-                4252822878758300859123897981450591353533073413197771768651442665752259397132,
-                6375614351688725206403948262868962793625744043794305715222011528459656738731
-            ],
-            [
-                21847035105528745403288232691147584728191162732299865338377159692350059136679,
-                10505242626370262277552901082094356697409835680220590971873171140371331206856
-            ]
-        );
-        vk.gamma2 = Pairing.G2Point(
-            [
-                11559732032986387107991004021392285783925812861821192530917403151452391805634,
-                10857046999023057135944570762232829481370756359578518086990519993285655852781
-            ],
-            [
-                4082367875863433681332203403145435568316851327593401208105741076214120093531,
-                8495653923123431417604973247489272438418190587263600148770280649306958101930
-            ]
-        );
-        vk.delta2 = Pairing.G2Point(
-            [
-                7398507350689558813165340153029711718531339594530713382535952511400220563796,
-                1506012156639833252275989782489414790712081632800342473254083004912542674274
-            ],
-            [
-                4031677550618857986830914063418669372844655023558685089932409841302325076791,
-                1593477473510602753678739765915618087350589354501364366559921979230657836188
-            ]
-        );
-        vk.IC = new Pairing.G1Point[](5);
-
-        vk.IC[0] = Pairing.G1Point(
-            7657453013080600726078477537563353687065524019804870066799834008389861407607,
-            8685901164953056600635317925027970572415723347755279600050476638006925034076
-        );
-
-        vk.IC[1] = Pairing.G1Point(
-            328347287025020767974405630803846918104736612380726342059575304769291270786,
-            17659234291197731341197163219056358146267731752704552764586696913333856517830
-        );
-
-        vk.IC[2] = Pairing.G1Point(
-            21032083054530545144494333246335839293914039387842369594492729566988213432517,
-            1536433841005546413961097766679183973551918073760154675490930443164981376988
-        );
-
-        vk.IC[3] = Pairing.G1Point(
-            186401485499102920312703913827541615727484862412358380427385584112776129997,
-            18922184230173269444971263978843170103443530679049082563930873209260975292851
-        );
-
-        vk.IC[4] = Pairing.G1Point(
-            12078731651748322692267053524070937817677298895078631349379335497351881125510,
-            890641629248220383806186525705126321607926178899369387826785425880517749306
-        );
-    }
-
-    function verify(uint256[] memory input, Proof memory proof) internal view returns (uint256) {
-        uint256 snark_scalar_field = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
-        VerifyingKey memory vk = verifyingKey();
-        require(input.length + 1 == vk.IC.length, "verifier-bad-input");
-        // Compute the linear combination vk_x
-        Pairing.G1Point memory vk_x = Pairing.G1Point(0, 0);
-        for (uint256 i = 0; i < input.length; i++) {
-            require(input[i] < snark_scalar_field, "verifier-gte-snark-scalar-field");
-            vk_x = Pairing.addition(vk_x, Pairing.scalar_mul(vk.IC[i + 1], input[i]));
-        }
-        vk_x = Pairing.addition(vk_x, vk.IC[0]);
-        if (
-            !Pairing.pairingProd4(
-                Pairing.negate(proof.A),
-                proof.B,
-                vk.alfa1,
-                vk.beta2,
-                vk_x,
-                vk.gamma2,
-                proof.C,
-                vk.delta2
-            )
-        ) return 1;
-        return 0;
-    }
-
-    /// @return r  bool true if proof is valid
-    function verifyProof(
-        uint256[2] memory a,
-        uint256[2][2] memory b,
-        uint256[2] memory c,
-        uint256[4] memory input
-    ) public view returns (bool r) {
-        // slither-disable-next-line uninitialized-local
-        Proof memory proof;
-        proof.A = Pairing.G1Point(a[0], a[1]);
-        proof.B = Pairing.G2Point([b[0][0], b[0][1]], [b[1][0], b[1][1]]);
-        proof.C = Pairing.G1Point(c[0], c[1]);
-        uint256[] memory inputValues = new uint256[](input.length);
-        for (uint256 i = 0; i < input.length; i++) {
-            inputValues[i] = input[i];
-        }
-        if (verify(inputValues, proof) == 0) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-}
