+++ test/state/stateV2.test.ts
@@ -1,7 +1,7 @@
 import { expect } from "chai";
 import { ethers } from "hardhat";
-import { publishState, publishStateWithStubProof } from "../utils/deploy-utils";
-import { StateDeployHelper } from "../../helpers/StateDeployHelper";
+import { publishState, publishStateWithStubProof } from "../utils/state-utils";
+import { DeployHelper } from "../../helpers/DeployHelper";
 import bigInt from "big-integer";
 
 const verifierStubName = "VerifierStub";
@@ -31,15 +31,13 @@ describe("State transition with real verifier", () => {
 
   before(async function () {
     this.timeout(5000);
-    const deployHelper = await StateDeployHelper.initialize();
+    const deployHelper = await DeployHelper.initialize();
     const contracts = await deployHelper.deployStateV2();
     state = contracts.state;
   });
 
   it("Zero-knowledge proof of state transition is not valid", async () => {
-    const modifiedStateTransition = JSON.parse(
-      JSON.stringify(stateTransitionsWithProofs[0])
-    );
+    const modifiedStateTransition = JSON.parse(JSON.stringify(stateTransitionsWithProofs[0]));
     modifiedStateTransition.pub_signals[2] = "100"; // change state to make zk proof invalid
 
     await expect(publishState(state, modifiedStateTransition)).to.be.revertedWith(
@@ -69,9 +67,7 @@ describe("State transition with real verifier", () => {
     expect(stInfoOld.id).to.be.equal(params.id);
     expect(stInfoOld.replacedByState).to.be.equal(params.newState);
     expect(stInfoOld.createdAtTimestamp).to.be.equal(0);
-    expect(stInfoOld.replacedAtTimestamp).to.be.equal(
-      stInfoNew.createdAtTimestamp
-    );
+    expect(stInfoOld.replacedAtTimestamp).to.be.equal(stInfoNew.createdAtTimestamp);
     expect(stInfoOld.createdAtBlock).to.be.equal(0);
     expect(stInfoOld.replacedAtBlock).to.be.equal(stInfoNew.createdAtBlock);
 
@@ -102,16 +98,10 @@ describe("State transition with real verifier", () => {
 
     expect(await state.stateExists(params.id, params.oldState)).to.be.equal(true);
     const stInfoOld = await state.getStateInfoByIdAndState(params.id, params.oldState);
-    expect(stInfoOld.replacedAtTimestamp).to.be.equal(
-      stInfoNew.createdAtTimestamp
-    );
-    expect(stInfoOld.createdAtTimestamp).to.be.equal(
-      stateInfoBeforeUpdate.createdAtTimestamp
-    );
+    expect(stInfoOld.replacedAtTimestamp).to.be.equal(stInfoNew.createdAtTimestamp);
+    expect(stInfoOld.createdAtTimestamp).to.be.equal(stateInfoBeforeUpdate.createdAtTimestamp);
     expect(stInfoOld.replacedAtBlock).to.be.equal(stInfoNew.createdAtBlock);
-    !expect(stInfoOld.createdAtBlock).to.be.equal(
-      stateInfoBeforeUpdate.createdAtBlock
-    );
+    !expect(stInfoOld.createdAtBlock).to.be.equal(stateInfoBeforeUpdate.createdAtBlock);
     expect(stInfoOld.id).to.be.equal(params.id);
     expect(stInfoOld.replacedByState).to.be.equal(params.newState);
 
@@ -125,7 +115,7 @@ describe("State transition negative cases", () => {
   let state;
 
   beforeEach(async () => {
-    const deployHelper = await StateDeployHelper.initialize();
+    const deployHelper = await DeployHelper.initialize();
     const contracts = await deployHelper.deployStateV2(verifierStubName);
     state = contracts.state;
   });
@@ -133,9 +123,7 @@ describe("State transition negative cases", () => {
   it("Old state does not match the latest state", async () => {
     await publishStateWithStubProof(state, stateTransitionsWithNoProofs[0]);
 
-    const modifiedStateTransition = JSON.parse(
-      JSON.stringify(stateTransitionsWithNoProofs[1])
-    );
+    const modifiedStateTransition = JSON.parse(JSON.stringify(stateTransitionsWithNoProofs[1]));
     modifiedStateTransition.oldState = 10;
 
     await expect(publishStateWithStubProof(state, modifiedStateTransition)).to.be.revertedWith(
@@ -146,9 +134,7 @@ describe("State transition negative cases", () => {
   it("Old state is genesis but identity already exists", async () => {
     await publishStateWithStubProof(state, stateTransitionsWithNoProofs[0]);
 
-    const modifiedStateTransition = JSON.parse(
-      JSON.stringify(stateTransitionsWithNoProofs[1])
-    );
+    const modifiedStateTransition = JSON.parse(JSON.stringify(stateTransitionsWithNoProofs[1]));
     modifiedStateTransition.isOldStateGenesis = true;
 
     await expect(publishStateWithStubProof(state, modifiedStateTransition)).to.be.revertedWith(
@@ -157,9 +143,7 @@ describe("State transition negative cases", () => {
   });
 
   it("Old state is not genesis but identity does not yet exist", async () => {
-    const modifiedStateTransition = JSON.parse(
-      JSON.stringify(stateTransitionsWithNoProofs[0])
-    );
+    const modifiedStateTransition = JSON.parse(JSON.stringify(stateTransitionsWithNoProofs[0]));
     modifiedStateTransition.isOldStateGenesis = false;
 
     await expect(publishStateWithStubProof(state, modifiedStateTransition)).to.be.revertedWith(
@@ -168,9 +152,7 @@ describe("State transition negative cases", () => {
   });
 
   it("ID should not be zero", async () => {
-    const modifiedStateTransition = JSON.parse(
-      JSON.stringify(stateTransitionsWithNoProofs[0])
-    );
+    const modifiedStateTransition = JSON.parse(JSON.stringify(stateTransitionsWithNoProofs[0]));
     modifiedStateTransition.id = 0;
 
     await expect(publishStateWithStubProof(state, modifiedStateTransition)).to.be.revertedWith(
@@ -179,9 +161,7 @@ describe("State transition negative cases", () => {
   });
 
   it("New state should not be zero", async () => {
-    const modifiedStateTransition = JSON.parse(
-      JSON.stringify(stateTransitionsWithNoProofs[0])
-    );
+    const modifiedStateTransition = JSON.parse(JSON.stringify(stateTransitionsWithNoProofs[0]));
     modifiedStateTransition.newState = 0;
 
     await expect(publishStateWithStubProof(state, modifiedStateTransition)).to.be.revertedWith(
@@ -200,9 +180,9 @@ describe("State transition negative cases", () => {
       isOldStateGenesis: false,
     };
 
-    await expect(
-      publishStateWithStubProof(state, stateTransition)
-    ).to.be.revertedWith("New state already exists");
+    await expect(publishStateWithStubProof(state, stateTransition)).to.be.revertedWith(
+      "New state already exists"
+    );
   });
 });
 
@@ -211,7 +191,7 @@ describe("StateInfo history", function () {
   let publishedStates: { [key: string]: string | number }[] = [];
 
   before(async () => {
-    const deployHelper = await StateDeployHelper.initialize();
+    const deployHelper = await DeployHelper.initialize();
     const contracts = await deployHelper.deployStateV2(verifierStubName);
     state = contracts.state;
 
@@ -227,11 +207,7 @@ describe("StateInfo history", function () {
 
     expect(stateHistoryLength).to.be.equal(stateTransitionsWithNoProofs.length + 1);
 
-    const stateInfos = await state.getStateInfoHistoryById(
-      id,
-      0,
-      stateHistoryLength
-    );
+    const stateInfos = await state.getStateInfoHistoryById(id, 0, stateHistoryLength);
     expect(stateInfos.length).to.be.equal(stateHistoryLength);
 
     const publishedState1 = publishedStates[0];
@@ -251,7 +227,7 @@ describe("GIST proofs", () => {
   let state: any;
 
   beforeEach(async () => {
-    const deployHelper = await StateDeployHelper.initialize();
+    const deployHelper = await DeployHelper.initialize();
     const contracts = await deployHelper.deployStateV2(verifierStubName);
     state = contracts.state;
   });
@@ -293,17 +269,11 @@ describe("GIST proofs", () => {
 
     const [root1info, root2info] = await state.getGISTRootHistory(1, 2);
 
-    const [r1] = await state.getGISTProofByTime(
-      id,
-      root1info.createdAtTimestamp
-    );
+    const [r1] = await state.getGISTProofByTime(id, root1info.createdAtTimestamp);
 
     expect(root1info.root).to.equal(r1);
 
-    const [r2] = await state.getGISTProofByTime(
-      id,
-      root2info.createdAtTimestamp
-    );
+    const [r2] = await state.getGISTProofByTime(id, root2info.createdAtTimestamp);
     expect(r2).to.equal(root2info.root);
   });
 
@@ -319,15 +289,9 @@ describe("GIST proofs", () => {
 
     const [root1info, root2info] = await state.getGISTRootHistory(1, 2);
 
-    const [root] = await state.getGISTProofByBlock(
-      id,
-      root1info.createdAtBlock
-    );
+    const [root] = await state.getGISTProofByBlock(id, root1info.createdAtBlock);
     expect(root1info.root).to.equal(root);
-    const [root2] = await state.getGISTProofByBlock(
-      id,
-      root2info.createdAtBlock
-    );
+    const [root2] = await state.getGISTProofByBlock(id, root2info.createdAtBlock);
     expect(root2info.root).to.equal(root2);
   });
 });
@@ -336,7 +300,7 @@ describe("GIST root history", () => {
   let state: any;
 
   beforeEach(async () => {
-    const deployHelper = await StateDeployHelper.initialize();
+    const deployHelper = await DeployHelper.initialize();
     const contracts = await deployHelper.deployStateV2(verifierStubName);
     state = contracts.state;
   });
@@ -352,15 +316,9 @@ describe("GIST root history", () => {
 
     const [rootInfo] = await state.getGISTRootHistory(1, 1);
 
-    const [rootB] = await state.getGISTProofByBlock(
-      id,
-      rootInfo.createdAtBlock
-    );
+    const [rootB] = await state.getGISTProofByBlock(id, rootInfo.createdAtBlock);
     expect(rootInfo.root).to.equal(rootB);
-    const [rootT] = await state.getGISTProofByTime(
-      id,
-      rootInfo.createdAtTimestamp
-    );
+    const [rootT] = await state.getGISTProofByTime(id, rootInfo.createdAtTimestamp);
     expect(rootInfo.root).to.equal(rootT).to.equal(rootB);
   });
 
@@ -369,10 +327,7 @@ describe("GIST root history", () => {
     const roots: any[] = [];
     const expRootInfos: any[] = [];
     for (const issuerStateJson of stateTransitionsWithNoProofs) {
-      const { blockNumber, timestamp } = await publishStateWithStubProof(
-        state,
-        issuerStateJson
-      );
+      const { blockNumber, timestamp } = await publishStateWithStubProof(state, issuerStateJson);
 
       const root = await state.getGISTRoot();
       roots.push(root);
@@ -395,22 +350,14 @@ describe("GIST root history", () => {
     const rootInfo0 = await state.getGISTRootInfo(roots[0]);
     const rootInfo1 = await state.getGISTRootInfo(roots[1]);
 
-    expect(rootInfo0.replacedAtTimestamp).to.equal(
-      expRootInfos[0].replacedAtTimestamp
-    );
-    expect(rootInfo0.createdAtTimestamp).to.equal(
-      expRootInfos[0].createdAtTimestamp
-    );
+    expect(rootInfo0.replacedAtTimestamp).to.equal(expRootInfos[0].replacedAtTimestamp);
+    expect(rootInfo0.createdAtTimestamp).to.equal(expRootInfos[0].createdAtTimestamp);
     expect(rootInfo0.replacedAtBlock).to.equal(expRootInfos[0].replacedAtBlock);
     expect(rootInfo0.createdAtBlock).to.equal(expRootInfos[0].createdAtBlock);
     expect(rootInfo0.replacedByRoot).to.equal(expRootInfos[0].replacedByRoot);
 
-    expect(rootInfo1.replacedAtTimestamp).to.equal(
-      expRootInfos[1].replacedAtTimestamp
-    );
-    expect(rootInfo1.createdAtTimestamp).to.equal(
-      expRootInfos[1].createdAtTimestamp
-    );
+    expect(rootInfo1.replacedAtTimestamp).to.equal(expRootInfos[1].replacedAtTimestamp);
+    expect(rootInfo1.createdAtTimestamp).to.equal(expRootInfos[1].createdAtTimestamp);
     expect(rootInfo1.replacedAtBlock).to.equal(expRootInfos[1].replacedAtBlock);
     expect(rootInfo1.createdAtBlock).to.equal(expRootInfos[1].createdAtBlock);
     expect(rootInfo1.replacedByRoot).to.equal(expRootInfos[1].replacedByRoot);
@@ -419,34 +366,38 @@ describe("GIST root history", () => {
 
 describe("Set Verifier", () => {
   it("Should set verifier", async () => {
-    const deployHelper = await StateDeployHelper.initialize();
+    const deployHelper = await DeployHelper.initialize();
     const { state, verifier } = await deployHelper.deployStateV2();
 
     const verifierAddress = await state.getVerifier();
     expect(verifierAddress).to.equal(verifier.address);
 
-    const newVerifierAddress = ethers.utils.getAddress("0x8ba1f109551bd432803012645ac136ddd64dba72");
+    const newVerifierAddress = ethers.utils.getAddress(
+      "0x8ba1f109551bd432803012645ac136ddd64dba72"
+    );
     await state.setVerifier(newVerifierAddress);
     const verifierAddress2 = await state.getVerifier();
     expect(verifierAddress2).to.equal(newVerifierAddress);
   });
 
   it("Should not set verifier if not owner", async () => {
-    const deployHelper = await StateDeployHelper.initialize();
+    const deployHelper = await DeployHelper.initialize();
     const { state, verifier } = await deployHelper.deployStateV2();
 
     const verifierAddress = await state.getVerifier();
     expect(verifierAddress).to.equal(verifier.address);
 
     const notOwner = (await ethers.getSigners())[1];
-    const newVerifierAddress = ethers.utils.getAddress("0x8ba1f109551bd432803012645ac136ddd64dba72");
+    const newVerifierAddress = ethers.utils.getAddress(
+      "0x8ba1f109551bd432803012645ac136ddd64dba72"
+    );
     await expect(state.connect(notOwner).setVerifier(newVerifierAddress)).to.be.revertedWith(
       "Ownable: caller is not the owner"
     );
   });
 
   it("Should allow verifier zero address to block any state transition", async () => {
-    const deployHelper = await StateDeployHelper.initialize();
+    const deployHelper = await DeployHelper.initialize();
     const { state } = await deployHelper.deployStateV2();
 
     await state.setVerifier(ethers.constants.AddressZero);
