--- test/vectorsGen/go.mod
@@ -1,12 +1,0 @@
-module vectorsGen
-
-go 1.13
-
-// replace github.com/iden3/go-iden3-crypto => ../../../go-iden3-crypto
-// replace github.com/iden3/go-iden3-core => ../../../go-iden3-core
-
-require (
-	github.com/iden3/go-iden3-core v0.0.8
-	github.com/iden3/go-iden3-crypto v0.0.5
-	github.com/stretchr/testify v1.5.1
-)
